Create separate function to buffer events
